A prop gaze toolbar
The size of the toolbar and the button is relative to the resolution.
Toolbar is a notify icon on the  taskbar. Toolbar will auto start when
user turn on the computer if user set it on, Separate the main function
bar and settings. All the button on toolbar is now clickable